Allow using a cache size of up to 4 TB. Obviously anything close to 4 GB will always fail on 32 bit systems.
git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@34792 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ stream/stream.h
@@ -191,7 +191,7 @@ int stream_seek_long(stream_t *s, off_t pos);
 void stream_capture_do(stream_t *s);
 
 #ifdef CONFIG_STREAM_CACHE
-int stream_enable_cache(stream_t *stream,int size,int min,int prefill);
+int stream_enable_cache(stream_t *stream,int64_t size,int64_t min,int64_t prefill);
 int cache_stream_fill_buffer(stream_t *s);
 int cache_stream_seek_long(stream_t *s,off_t pos);
 #else
